Refactor configurations into one file
All the configs files have been refactored into one consolidated
configuration file. This allows for quicker additions as well as
extensions. The interface to the rest of the system is also quicker and
easier since there is just one "config" object that is the source of
truth.

All references have been updated but the billing model was not
thoroughly tested.